Moving torpedo methods to common
